+++ src/main/java/com/studypals/domain/fileManage/ObjectStorage.java
@@ -1,4 +1,4 @@
-package com.studypals.global.file;
+package com.studypals.domain.fileManage;
 
 /**
  * Object Storage 의 인터페이스입니다. 메서드를 정의합니다.
